Make sure amount has two decimal places
SmartDebit expects to receive amounts in pennies. Or if there's a decimal in the amount it's assumed to be pounds, and it's then converted to pennies. The amount calculated from the linked recurs risks being in pounds, but with no decimal places - so gets interpreted as pennies. Safest is to always pass through a value with two decimal places.